Fix some bugs

+++ i8_terminal/common/metrics.py
@@ -28,7 +28,7 @@ def get_indicators_list(indicator: Optional[str] = None) -> List[str]:
 
 
 def find_similar_fin_metric(metric: str) -> Optional[str]:
-    metrics_meta_data = get_all_metrics_df()[["metric_name", "name"]]
+    metrics_meta_data = get_all_metrics_df()[["metric_name"]]
     best_match_similarity = 0.0
     best_match = ""
     for idx, m in metrics_meta_data.iterrows():
@@ -67,8 +67,20 @@ def get_all_metrics_df() -> DataFrame:
     else:
         all_metrics = MetricsApi().get_list_metrics_metadata(page_size=1000)
         df = DataFrame([m.to_dict() for m in all_metrics])
-        df["categories"] = [",".join(str(cat)) for cat in df["categories"]]
+        df["categories"] = [str(cat) for cat in df["categories"]]
         df = df.drop(columns=["id", "last_modified"])
+        df.to_csv(metric_path, index=False)
+
+    return df
+
+
+def get_all_financial_metrics_df() -> DataFrame:
+    metric_path = f"{SETTINGS_FOLDER}/financial_metrics_metadata.csv"
+    if os.path.exists(metric_path) and not is_cached_file_expired(metric_path):
+        df = read_csv(metric_path)
+    else:
+        all_metrics = MetricsApi().get_list_financial_metrics_metadata()
+        df = DataFrame([m.to_dict() for m in all_metrics])
         df.to_csv(metric_path, index=False)
 
     return df
